tasks finalised

+++ controllers/contacts.js
@@ -4,18 +4,28 @@ const { HttpError, ctrlWrapper } = require("../helpers");
 
 const getAll = async (req, res) => {
   const { _id: owner } = req.user;
-  const { page = 1, limit = 12 } = req.query;
+  const { page = 1, limit = 20, favorite } = req.query;
   const skip = (page - 1) * limit;
-  const result = await Contact.find({ owner }, "-createdAt -updatedAt", {
-    skip,
-    limit,
-  }).populate("owner", "email");
+
+  const result = await Contact.find(
+    favorite ? { owner, favorite } : { owner },
+    "-createdAt -updatedAt",
+    {
+      skip,
+      limit,
+    }
+  ).populate("owner", "email subscription");
+
   res.json(result);
 };
 
 const getById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { id } = req.params;
-  const result = await Contact.findById(id);
+  const result = await Contact.findById(
+    { _id: id, owner },
+    "-createdAt -updatedAt"
+  ).populate("owner", "email subscription");
   if (!result) {
     throw HttpError(404, `Contact with ${id} not found`);
   }
@@ -24,13 +34,16 @@ const getById = async (req, res) => {
 
 const add = async (req, res) => {
   const { _id: owner } = req.user;
-  const result = await Contact.create(...req.body, owner);
+  const result = await Contact.create({ ...req.body, owner });
   res.status(201).json(result);
 };
 
 const updateById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { id } = req.params;
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  const result = await Contact.findByIdAndUpdate({ _id: id, owner }, req.body, {
+    new: true,
+  }).populate("owner", "email subscription");
   if (!result) {
     throw HttpError(404, `Contact with ${id} not found`);
   }
@@ -38,9 +51,11 @@ const updateById = async (req, res) => {
 };
 
 const updateFavorite = async (req, res) => {
-  console.log(req.body);
+  const { _id: owner } = req.user;
   const { id } = req.params;
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  const result = await Contact.findByIdAndUpdate({ _id: id, owner }, req.body, {
+    new: true,
+  }).populate("owner", "email subscription");
   if (!result) {
     throw HttpError(404, `Contact with ${id} not found`);
   }
@@ -48,8 +63,12 @@ const updateFavorite = async (req, res) => {
 };
 
 const deleteById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { id } = req.params;
-  const result = await Contact.findByIdAndRemove(id);
+  const result = await Contact.findByIdAndRemove({ _id: id, owner }).populate(
+    "owner",
+    "email subscription"
+  );
   if (!result) {
     throw HttpError(404, `Contact with id ${id} not found`);
   }
